moved -all functions in erc20 facade
