+++ scripts/integrationTests/smokeTest1/auctionStart.ts
@@ -19,7 +19,9 @@ export const smokeTest = async ({
   ethFacade,
   constants,
   getLPUnlockedBalanceAll,
-  getLPLockedBalanceAll
+  getLPLockedBalanceAll,
+  getBalancesAll,
+  approveAll
 }: TestRunner) => {
   const optionRoundFacade = await getOptionRoundFacade(
     provider,
@@ -80,7 +82,7 @@ export const smokeTest = async ({
       spender: optionRoundFacade.optionRoundContract.address,
     },
   ];
-  await ethFacade.approveAll(approveAllData);
+  await approveAll(approveAllData);
   await mineNextBlock(provider.channel.nodeUrl);
 
   //Place bids according to story script
@@ -88,7 +90,7 @@ export const smokeTest = async ({
   const totalOptionAvailable =
     await optionRoundFacade.getTotalOptionsAvailable();
 
-  const ethBalancesBefore = await ethFacade.getBalancesAll(
+  const ethBalancesBefore = await getBalancesAll(
     optionBidderAccounts
   );
 
@@ -111,7 +113,7 @@ export const smokeTest = async ({
   ];
   await optionRoundFacade.placeBidsAll(placeBidsData);
 
-  const ethBalancesAfter = await ethFacade.getBalancesAll(optionBidderAccounts);
+  const ethBalancesAfter = await getBalancesAll(optionBidderAccounts);
 
   const bidArrays = await optionRoundFacade.getBidsForAll(optionBidderAccounts);
